requirements.txt
fastapi
python-multipart
python-jose[cryptography]
passlib[bcrypt]
uvicorn
alembic
python-dotenv

